now return float

+++ main.py
@@ -10,10 +10,9 @@ if __name__ == '__main__':
     n = 10
     data = ext.ExtractData(sites_file, onshore_file, offshore_file)
     data.capacities()
-    z,x = optimize(data.onshore_capacities().astype(np.float64), 
-             data.offshore_capacities().astype(np.float64), 
-             data.onshore_rendements().astype(np.float64), 
-             data.offshore_rendements().astype(np.float64), 
-             ## hardcodé le cast des strings en floats, sans doute mieux de faire ça dans extract_data
+    z,x = optimize(data.onshore_capacities(),
+             data.offshore_capacities(),
+             data.onshore_rendements(),
+             data.offshore_rendements(),
              P=10000, k=0.4, 
              checkpoints=True)
